+++ packages/frameworks/react/src/date-picker/date-picker-root.tsx
@@ -51,6 +51,7 @@ export const DatePickerRoot = forwardRef<HTMLDivElement, DatePickerRootProps>((p
       'onValueChange',
       'onViewChange',
       'open',
+      'open.controlled',
       'parse',
       'positioning',
       'readOnly',