+++ packages/frameworks/react/src/select/select-root.tsx
@@ -55,6 +55,7 @@ const SelectImpl = <T extends CollectionItem>(
       'onPointerDownOutside',
       'onValueChange',
       'open',
+      'open.controlled',
       'positioning',
       'readOnly',
       'selectOnBlur',