+++ packages/frameworks/solid/src/date-picker/date-picker-root.tsx
@@ -44,6 +44,7 @@ export const DatePickerRoot: ArkComponent<'div', ElementProps> = (props: DatePic
     'onValueChange',
     'onViewChange',
     'open',
+    'open.controlled',
     'parse',
     'positioning',
     'readOnly',