+++ packages/frameworks/solid/src/dialog/dialog-root.tsx
@@ -35,6 +35,7 @@ export const DialogRoot = (props: DialogRootProps) => {
     'onOpenChange',
     'onPointerDownOutside',
     'open',
+    'open.controlled',
     'preventScroll',
     'restoreFocus',
     'role',