+++ packages/frameworks/vue/src/date-picker/date-picker.props.ts
@@ -53,6 +53,10 @@ export const props = {
   open: {
     type: Boolean as PropType<Context['open']>,
   },
+  'open.controlled': {
+    type: Boolean as PropType<Context['open.controlled']>,
+    default: undefined,
+  },
   positioning: {
     type: Object as PropType<Context['positioning']>,
   },