+++ packages/frameworks/vue/src/select/select.props.ts
@@ -48,6 +48,10 @@ export const props = {
     type: Boolean as PropType<Context['open']>,
     default: undefined,
   },
+  'open.controlled': {
+    type: Boolean as PropType<Context['open.controlled']>,
+    default: undefined,
+  },
   positioning: {
     type: Object as PropType<Context['positioning']>,
   },